Отправка отчета

+++ src/main/java/pro/sky/telegrambot/service/PhotoService.java
@@ -26,7 +26,7 @@ public class PhotoService {
         this.petService = petService;
         this.photoPetRepository = photoPetRepository;
     }
-    public void uploadPhoto(Long petId, MultipartFile file) throws IOException {
+    public void uploadPhoto(int petId, MultipartFile file) throws IOException {
         Pet pet = petService.findPet(petId);
 
         Path filePath = Path.of(coversDir, petId + "." + getExtension(file.getOriginalFilename()));
@@ -55,7 +55,7 @@ public class PhotoService {
         return fileName.substring(fileName.lastIndexOf(".") + 1);
     }
 
-    public Photo findPhoto(Long Id) {
+    public Photo findPhoto(int Id) {
         return photoPetRepository.findPhotoById(Id).orElse(new Photo());
     }
